refactor code to isolate elasticsearch dependencies
