+++ src/main/java/gov/nasa/pds/registry/common/connection/ClientConstants.java
@@ -1,11 +1,11 @@
-package gov.nasa.pds.registry.common.connection.es;
+package gov.nasa.pds.registry.common.connection;
 
 /**
  * Constants used by different classes.
  * 
  * @author karpenko
  */
-public class ClientConstants
+class ClientConstants
 {
     public static final String AUTH_TRUST_SELF_SIGNED = "trust.self-signed";
 }
