+++ src/main/java/gov/nasa/pds/registry/common/connection/es/EsClientFactory.java
@@ -10,7 +10,7 @@ import gov.nasa.pds.registry.common.ConnectionFactory;
  * 
  * @author karpenko
  */
-public class EsClientFactory
+class EsClientFactory
 {
     /**
      * Create Elasticsearch rest client.
