+++ src/main/java/gov/nasa/pds/registry/common/connection/es/EsUtils.java
@@ -2,21 +2,18 @@ package gov.nasa.pds.registry.common.connection.es;
 
 
 import java.util.List;
-import java.util.Map;
 
 import org.apache.http.HttpHost;
 import org.elasticsearch.client.Response;
 import org.elasticsearch.client.ResponseException;
-
-import com.google.gson.Gson;
-
+import gov.nasa.pds.registry.common.util.SearchResponseParser;
 
 /**
  * Elasticsearch utility methods.
  * 
  * @author karpenko
  */
-public class EsUtils
+class EsUtils
 {
     /**
      * Parse Elasticsearch URL
@@ -99,46 +96,13 @@ public class EsUtils
         String lines[] = msg.split("\n");
         if(lines.length < 2) return msg;
         
-        String reason = extractReasonFromJson(lines[1]);
+        String reason = SearchResponseParser.extractReasonFromJson(lines[1]);
         if(reason == null) return msg;
         
         return reason;
     }
     
     
-    /**
-     * Extract error message from Elasticsearch response JSON.
-     * @param json response JSON
-     * @return error message
-     */
-    @SuppressWarnings("rawtypes")
-    public static String extractReasonFromJson(String json)
-    {
-        try
-        {
-            Gson gson = new Gson();
-            Object obj = gson.fromJson(json, Object.class);
-            
-            obj = ((Map)obj).get("error");
-            
-            Object rc = ((Map)obj).get("root_cause");
-            if(rc != null)
-            {
-                List list = (List)rc;
-                obj = ((Map)list.get(0)).get("reason");
-            }
-            else
-            {
-                obj = ((Map)obj).get("reason");
-            }
-            
-            return obj.toString();
-        }
-        catch(Exception ex)
-        {
-            return null;
-        }
-    }
 
     
     /**
