+++ src/main/java/gov/nasa/pds/registry/common/util/SearchResponseParser.java
@@ -1,8 +1,9 @@
-package gov.nasa.pds.registry.common.connection.es;
+package gov.nasa.pds.registry.common.util;
 
 import java.io.InputStream;
 import java.io.InputStreamReader;
-
+import java.util.List;
+import java.util.Map;
 import org.elasticsearch.client.Response;
 
 import com.google.gson.Gson;
@@ -172,4 +173,37 @@ public class SearchResponseParser
         cb.onRecord(lastId, src);
     }    
 
+    /**
+     * Extract error message from search response JSON.
+     * @param json response JSON
+     * @return error message
+     */
+    @SuppressWarnings("rawtypes")
+    public static String extractReasonFromJson(String json)
+    {
+        try
+        {
+            Gson gson = new Gson();
+            Object obj = gson.fromJson(json, Object.class);
+            
+            obj = ((Map)obj).get("error");
+            
+            Object rc = ((Map)obj).get("root_cause");
+            if(rc != null)
+            {
+                List list = (List)rc;
+                obj = ((Map)list.get(0)).get("reason");
+            }
+            else
+            {
+                obj = ((Map)obj).get("reason");
+            }
+            
+            return obj.toString();
+        }
+        catch(Exception ex)
+        {
+            return null;
+        }
+    }
 }
